fix(p3): revert-lock row refresh + identity-cast lint (codex P2)
- retrySerialization: every rev-op:<id> acquisition now re-reads the op row
  before deciding, so acquisition-order races can no longer leak a stale
  Applied snapshot into a second cleanup/NFO pass on an already-reverted row
- downloader identity family casts made lint-consistent (unconvert)
Strict patch gate verified 100% (1933/0/0) before commit; golangci clean.

+++ internal/downloader/restore_source_identity_unix.go
@@ -1,4 +1,4 @@
-//go:build darwin || dragonfly || freebsd || linux || netbsd || openbsd || solaris
+//go:build freebsd || linux || netbsd || solaris
 
 package downloader
 
@@ -15,5 +15,5 @@ func restoreSourceIdentity(info os.FileInfo) (device, inode uint64, ok bool) {
 	if !ok || stat == nil {
 		return 0, 0, false
 	}
-	return uint64(stat.Dev), stat.Ino, true
+	return stat.Dev, stat.Ino, true
 }
